% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpg.R
\name{bpg}
\alias{bpg}
\title{Batch process for glucometrics}
\usage{
bpg(dat, span.months = 1, hypocutoffs, hypercutoffs, normalrange)
}
\arguments{
\item{dat, }{as in \code{\link{GenGluM}}.}

\item{span.months, }{as in \code{\link{GenGluM}}.}

\item{hypocutoffs, }{as in \code{\link{GenGluM}}.}

\item{hypercutoffs, }{as in \code{\link{GenGluM}}.}

\item{normalrange, }{as in \code{\link{GenGluM}}.}
}
\description{
Internal function
}
\author{
Ying Chen, Yilin Ning
}
