% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFPlot.R
\name{CDFPlot}
\alias{CDFPlot}
\title{Cumulative distribution function plot}
\usage{
CDFPlot(dat)
}
\arguments{
\item{dat, }{as in \code{\link{PerformExclusion}}.}
}
\description{
Plot the cumulative distribution of daily blood glucose timings
}
\author{
Ying Chen, Chuen Seng Tan
}
