% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPlot_function.R
\name{classPlot}
\alias{classPlot}
\title{This does a scatterplot according to the classifications}
\usage{
classPlot(
  data,
  Xvar,
  Yvar,
  classVar,
  classes = NULL,
  legPos = "bottomright",
  leg.cex = 0.8,
  col.vec = c(2, 3, 4, 6, 1, 7),
  pch.vec = c(1, 2, 3, 4, 6, 9),
  ...
)
}
\arguments{
\item{data}{is the matrix of data to use}

\item{Xvar}{is variable from data to be plotted on x axis}

\item{Yvar}{is variable from data to be plotted on y axis}

\item{classVar}{is variable to plot by}

\item{classes}{are the specific classes to plot (default to all)}

\item{legPos}{is place for legend (default "bottomright")}

\item{col.vec}{and pch.vec are for points on plot}
}
\value{
A plot
}
\description{
This does a scatterplot according to the classifications
}
\author{
Brett Amidan
}
