% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels_function.R
\name{createModels}
\alias{createModels}
\title{This function performs Logistic Regression on the Qc Data}
\usage{
createModels(opt.lambda, cotdata, outputFolder, smaq.metrics, q.metrics)
}
\arguments{
\item{opt.lambda}{is the optimal or desired lambda for the LLRC}

\item{cotdata}{is the curated dataset}

\item{outputFolder}{is the path to store results}

\item{smaq.metrics}{are the smaq metrics}

\item{q.metrics}{are the q metrics}
}
\value{
LLRC models for each equipment
}
\description{
This function performs Logistic Regression on the Qc Data
}
\author{
Brett Amidan
}
