% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVlogisticReg2_function.R
\name{CVlogisticReg2}
\alias{CVlogisticReg2}
\title{CV Logistic Regression}
\usage{
CVlogisticReg2(the.data, respName, cutoff = 0.01)
}
\arguments{
\item{the.data}{is the matrix of data to use}

\item{respName}{is the Y variable to be regressed on}

\item{cutoff}{is the desired LASSO parameter (lambda)}
}
\value{
A list with coefficients, predictions, actuals
}
\description{
This function performs the cross validation (80/20) LLRC
}
\author{
Brett Amidan
}
