% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataScrubbing.R
\name{DataScrubbing}
\alias{DataScrubbing}
\title{Data scrubbing}
\usage{
DataScrubbing(dat, unitVal)
}
\arguments{
\item{dat, }{the output object of \code{\link{FormatDate}}.}

\item{unitVal, }{the unit indicator, 1 for mmol/L, 2 for md/dL.}
}
\value{
The function will return a list of three elements. The first element will be the data after data scrubbing. 
The second element will be a vector of counts for the situations that decribed in the first three steps taken to clean the data.
The last elemenet will report the data containing repeated measurements.
}
\description{
If the unit of BG measurement is mmol/L, this function, performs the following steps: 1. "<0.6" or ">33.3" reset to 0.6 and 33.3 respectively
2. "<" or ">" for middle values, decrement or increment by 0.1 3 "Low/lo" or "H/High" reset to limiting values 0.6 and 33.3
4. take the average BG at the same timing for particular patient, this might be system error
Report affected proportions of readings.
}
\author{
Ying Chen
}
