% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatDate.R
\name{FormatDate}
\alias{FormatDate}
\title{Format date variables using \code{\link{recogniseDateTime}}}
\usage{
FormatDate(dat, yy, mm)
}
\arguments{
\item{dat}{A \emph{data.frame}, has at least one key variable:
 \itemize{
\item{RESULT.DATE} an object of classes have information calender date and times
}}

\item{yy}{A numeric value specifies the year of the data set.}

\item{mm}{A numeric value specifies the month of the data set.}
}
\value{
A \emph{data.table}, has at least one key variable:
\itemize{
\item{RESULT.DATE} an object of classes \dQuote{POSIXlt} and \dQuote{POSIXct}
  representing calender date and times.
}
}
\description{
Format date variable from character to R class \dQuote{POSIXct}
  so that we can manipulate with date variables
}
\examples{
# Load example data
data("gluDat")
head(gluDat)
gluDat2 <- FormatDate(dat = gluDat, yy = 2016, mm = 7)
head(gluDat2)
}
\author{
Ying Chen
}
