% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEpisode.R
\name{GenEpisode}
\alias{GenEpisode}
\title{Calculate pseudo admission episodes}
\usage{
GenEpisode(dat, epiMethod = "Admininfo")
}
\arguments{
\item{dat}{A \emph{data.frame} with at least have three key variables:
\itemize{
 \item{RESULT.DATE} a date and time column of classes \dQuote{POSIXlt} and
   \dQuote{POSIXct} (can be generated by \code{\link{GenGluM}});
 \item{ADMISSION.ID} a character or numeric column for the ID of each
   hospital stay;
 \item{RESULT} a numeric column for Blood Glucose readings.
}}

\item{epiMethod}{This indicates the method takes to compute episodes. If
\code{Admininfo} (the default), the admission id will be assumed to be case
id; if \code{Pseudo}, then episode number will be generated using 48 hours
as a cutoff.}
}
\value{
The input \emph{data.table} with the following additional variables:
\item{LOS.EACH}{time difference between readings}
\item{LOS.PSUM}{cumulative time differences between readings}
\item{EPISODE.ID}{In the case of pseudo, admission times will be increase by
1 if LOS.EACH greater than 48 hours. Otherwise, it will always be 1,
corresponding to that each admission id identifies one hospital stay.}
\item{mond}{day of month}
\item{weekd}{day of week}
\item{yday}{day of year}
\item{hour}{numerical hours}
\item{weekn}{number of weeks in a year}
These variables are created using original key variable \code{RESULT.DATE}.
}
\description{
Read in glucose data and calculate pseudo admission episodes,
  calculate the time interval between consecutive readings and accumulated
  time intervals within each episode for each individual, identify the day in
  a week, month, year, week number in a year, convert timings into hours
  using the standard 24-hour clock.
}
\examples{
# Load example data
data("gluDat")
# Process date-time variable and blood glucose readings:
gluDat2 <- FormatDate(dat = gluDat, yy = 2016, mm = 7)
gluDat3_ls <- DataScrubbing(dat = gluDat2, unitVal = 1)
# Specify admission episodes based on admission ID:
gluDat4 <- GenEpisode(dat = gluDat3_ls$dat, epiMethod = "Admininfo")
}
\author{
Chuen Seng Tan, Ying Chen
}
