% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGluM.R
\name{GenGluM}
\alias{GenGluM}
\title{Generate glucometrics}
\usage{
GenGluM(dat, hypocutoffs, hypercutoffs, normalrange, hgicutoff, unitVal)
}
\arguments{
\item{dat}{A \emph{data.table} prepared by \code{GenEpisode}}

\item{hypocutoffs}{A vector of numeric values indicating the recommened
cutoffs for hypoglycemia, from mild, moderate to severe hypoglycemia. Hypoglycemia is defined as less than the specified cutoff value.}

\item{hypercutoffs}{A vector of numeric values indicating the recommened
cutoffs for hyperglycemia, from mild, moderate to severe hyperglycemia. Hyperglycemia is defined as no less than the specified cutoff value.}

\item{normalrange}{A vector of numeric values indicating the recommened range
for normal glycemia, where the first value is the lower range, and the
second value is the upper range. A BG is considered in recommended range if it is no less than a lower bound and less than a upper bound at the same time. 
@param hgicutoff A numeric value indicating the cutoff used for calculating hyperglycemic index (HGI).
@param unitVal A unit indicator. 1 stands for mmol/L, 2 stands for md/dL}
}
\value{
A list having three dataframes
\itemize{
\item{popstat} A vector of all the indices at population level
\item{patientdaystat} A vector of all the indices at patient day level
\item{peradmissionstat} A vector of all the indices at admission level
}
}
\description{
Generate glucometrics for specific ward/wards during specified time period
}
\author{
Ying Chen, Chuen Seng Tan
}
