% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenHypoIntervals.R
\name{GenHypoIntervals}
\alias{GenHypoIntervals}
\title{Calculate hypoglycemia management time intervals}
\usage{
GenHypoIntervals(dat)
}
\arguments{
\item{dat}{A \emph{data.table} prepared by \code{\link{DataScrubbing}}}
}
\value{
A \emph{list} contains the time intervals.
}
\description{
There are three types of intervals: hypo to next documented BG, hypo to resoultion of hypo, and resolution of hypo to next documented hypo. This function will calculate these intervals.
}
\author{
Ying Chen
}
