% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gluDat}
\alias{gluDat}
\title{Example glucose data}
\format{
A data frame with 4459 rows and 4 variables:
\describe{
  \item{ADMISSION.ID}{Randomly generated integers indicating each patient-stay.}
  \item{RESULT}{Randomly generated glucose readings (mmol/L). 6 of the 4459
    entries are non-numeric by design.}
  \item{RESULT.DATE}{Date-time stamp of glucose readings, randomly generated
    to be within the calendar month of July 2020.}
  \item{LOCATION}{Randomly generated strings indicating 45 different ward
    locations.}
}
}
\usage{
gluDat
}
\description{
Simulated glucose data in July 2020.
}
\keyword{datasets}
