% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticReg_function.R
\name{logisticReg}
\alias{logisticReg}
\title{This function performs Logistic Regression}
\usage{
logisticReg(data, respName, cutoff = 0.01)
}
\arguments{
\item{data}{is the matrix of data to use}

\item{respName}{is the Y variable to be regressed on}

\item{cutoff}{is the desired LASSO parameter (lambda)}
}
\value{
A list with coefficients and formula
}
\description{
This function performs an individual LLRC and is called by CVlogisticReg2
}
\details{
This function requires glmnet package
}
\author{
Brett Amidan
}
