% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncuratedPrediction_function.R
\name{noncuratedPrediction}
\alias{noncuratedPrediction}
\title{This function performs prediction}
\usage{
noncuratedPrediction(ncdataFilename, modelsFile, dataFilename, outputFolder)
}
\arguments{
\item{ncdataFilename}{is full filename of the noncurated dataset}

\item{modelsFile}{is the full filename of the models file produced earlier}

\item{dataFilename}{is the full filename of the Rdata dataset}

\item{outputFolder}{is the path to store the output}

\item{instrument}{is the instrument to run the predictions for}
}
\value{
A vector containing the predictions
}
\description{
This function predicts the non-curated data
only works for selected instrument
}
\examples{
noncuratedPrediction(ncdataFilename="D:/Data/ncData.csv",modelsFile="D:/Results/Models_paper.Rdata",dataFilename="D:/Data/Data.Rdata",outputFolder="D:/Results/",instrument="VOrbitrap")
}
\author{
Brett Amidan
}
