% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCalfunc.R
\name{pCalfunc}
\alias{pCalfunc}
\title{Calculating KS p-values}
\usage{
pCalfunc(dat)
}
\arguments{
\item{dat}{A \emph{data.table} prepared by \code{datapreparation}}
}
\value{
A \emph{data.frame} with columns being
  \itemize{
    \item{day} The column indicates day counts
    \item{Bs.1ksPvalue} The bootstrapped One-Sample Kolmogorov-Smirnov test
    p-values
    \item{X1ksPval} The exact One-Sample Kolmogorov-Smirnov test p-values
    \item{X1ksStat} The test statistics of One-Sample Kolmogorov-Smirnov
    test
    \item{Bs.2ksPvalue} The bootstrapped Two-Sample Kolmogorov-Smirnov test
    p-values
    \item{X2ksPval} The exact Two-Sample Kolmogorov-Smirnov test p-values
    \item{X2ksStat} The test statistics of Two-Sample Kolmogorov-Smirnov
    testtrapped One-Sample Kolmogorov-Smirnov test p-values
    \item{X2lrPval} The exact Two-Sample Likilihood Ratio test p-values
    \item{NoOfReadings} The number of measurements taken within that day
  }
}
\description{
This function performs one-sample Kolmogorov-Smirnov test of the distribution of
daily BG timings against uniformation distribution, and two-sample Kolmogorov-Smirnov test for
the distributions of BG timings from two consecutive days, and provides p-values.
}
\author{
Tan Chuen Seng, Chen Ying
}
