% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformExclusion.R
\name{PerformExclusion}
\alias{PerformExclusion}
\title{Apply exclusion criterion to data}
\usage{
PerformExclusion(
  preDat,
  crtVec = c(crt.los = TRUE, crt.freq = TRUE, crt.1stday = TRUE),
  ageRange = c(16, 120),
  losNum_hour = 24,
  freqNum = 5
)
}
\arguments{
\item{preDat}{A \emph{data.table} which should be the output of
\code{\link{GenEpisode}}}

\item{crtVec}{A boolean vector of length 3, indicating whether to exclude
patients by length of stay, frequency of BG measurements, and whether to
exclude the readings taken within the first 24 hours.}

\item{ageRange}{A vector specifying the age range. Any patient younger than
the minimum age or older than the maximum age should be excluded. This
exclusion criterion requires the data to have a BIRTH.DATE column or an AGE
column, otherwise it will be ingored.}

\item{losNum_hour}{A number specifying the length of stay criterion. A hospital stay need to be 
greater than this value to be included in the analysis. This is an optional exclusion criterion.}

\item{freqNum}{A numeric value specifying the smallest frequency of BG within a hospital stay 
needed for the analysis. This is an optional criterion.}
}
\description{
By default, there are three criterion, age criterion, duration
  criterion, and frequency criterion.
}
\author{
Ying Chen
}
