% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQCdata_function.R
\name{readQCdata}
\alias{readQCdata}
\title{This function reads in the QC / Ratings Dataset}
\usage{
readQCdata(cotDataName, outDataName)
}
\arguments{
\item{cotDataName}{is the full filename of the curated dataset (csv)}

\item{outDataName}{is the full filename of the output Rdata object to
contain the resulting datasets}
}
\value{
An rdata object with the data objects needed for analyses
}
\description{
This function reads in the QC / Ratings Dataset
}
\examples{
readQCdata(cotDataName="D:/Data/Data.csv",outDataName="D:/Data/Data.Rdata")
}
\author{
Brett Amidan
}
