% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recogniseDateTime.R
\name{recog_one_date}
\alias{recog_one_date}
\title{Recognise one date given its month and year}
\usage{
recog_one_date(date, month, year)
}
\arguments{
\item{date}{a string of date, where date, month and year are separated by
either "-" or "/", but not both.}

\item{month}{a number (1 to 12) indicating the month this time point is in.}

\item{year}{a number (with 4 digits) indicating the year this time point is in.}
}
\value{
Returns the date as a character variable of the format %d/%m/%Y.
}
\description{
Recognise one date given its month and year
}
