% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recogniseDate.R
\name{recogniseDate}
\alias{recogniseDate}
\title{Recognise date}
\usage{
recogniseDate(date, month, year)
}
\arguments{
\item{date}{a single time point satisfying the assumptions, see details.}

\item{month}{a number (1 to 12) indicating the month this time point is in}

\item{year}{a number (00 to 99) indicating the year this time point is in}
}
\value{
Returns a character variable of the format %d/%m/%y %H:%M:%S
}
\description{
Recognise the date component of a time variable
}
\details{
none
}
\author{
Yilin Ning, Ying Chen
}
