% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLLRC_function.R
\name{trainingLLRC}
\alias{trainingLLRC}
\title{This is the Training function}
\usage{
trainingLLRC(dataFilename, outputFolder, optKappa = 5)
}
\arguments{
\item{dataFilename}{is full filename of the Rdata object containing the data}

\item{outputFolder}{is the path to store the output}

\item{optKappa}{is the kappa to run the analyses on (loss function value)}
}
\value{
A list containing the models used, the results (sensitivity etc),
univariate.results (VOrbitrap vs single variables), optimal values of
lambda, tau, and kappa for each equipment
}
\description{
This function uses LASSO and Logistic regression and cross validation
to determine the optimal lambda (lasso parameter) and
tau (cutoff for kappa cost function)
}
\examples{
trainingLLRC(dataFilename="D:/Data/Data.Rdata",outputFolder="D:/Results/",optKappa=5)
}
\author{
Brett Amidan
}
