% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGluM.R
\name{bpg}
\alias{bpg}
\title{Batch process for glucometrics}
\usage{
bpg(
  dat,
  span.months = 1,
  hypocutoffs,
  hypercutoffs,
  normalrange,
  hgicutoff,
  unitVal
)
}
\arguments{
\item{dat}{A \emph{data.table} prepared by \code{\link{GenEpisode}}.}

\item{span.months}{The number of months to aggregate by in the report, e.g.,
\code{span.months = 1} for monthly report, \code{span.months = 3} for
quarterly report and \code{span.months = 6} for half-annually report.}

\item{hypocutoffs}{A vector of numeric values indicating the recommended
cutoffs for hypoglycemia, from mild, moderate to severe hypoglycemia.
Hypoglycemia is defined as less than the specified cutoff value.}

\item{hypercutoffs}{A vector of numeric values indicating the recommended
cutoffs for hyperglycemia, from mild, moderate to severe hyperglycemia.
Hyperglycemia is defined as no less than the specified cutoff value.}

\item{normalrange}{A vector of numeric values indicating the recommended range
for normal glycemia, where the first value is the lower range, and the
second value is the upper range. A BG is considered in recommended range if
it is no less than a lower bound and less than a upper bound at the same
time.}

\item{hgicutoff}{A numeric value indicating the cutoff used for calculating
hyperglycemic index (HGI).}

\item{unitVal}{A unit indicator. 1 stands for mmol/L, 2 stands for md/dL}
}
\value{
Returns a list, where each item is an output from \code{\link{GenGluM}}.
}
\description{
Internal function
}
\seealso{
\code{\link{GenGluM}}
}
\author{
Ying Chen, Yilin Ning
}
