% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataScrubbing.R
\name{DataScrubbing}
\alias{DataScrubbing}
\title{Data scrubbing}
\usage{
DataScrubbing(dat, unitVal)
}
\arguments{
\item{dat}{A \emph{data.frame} that at least have the following key variables:
\itemize{
 \item{RESULT} a character or numeric column for Blood Glucose readings;
 \item{RESULT.DATE} a date and time column of classes \dQuote{POSIXlt} and
   \dQuote{POSIXct} (can be generated by \code{\link{GenGluM}});
 \item{ADMISSION.ID} a character or numeric column for the ID of each
   hospital stay;
 \item{LOCATION} a character column for the ward of each hospital stay.
}}

\item{unitVal}{The unit indicator, 1 for mmol/L, 2 for md/dL.}
}
\value{
The function will return a list of three elements. The first element
  will be the data after data scrubbing. The second element will be a vector
  of counts for the situations described in the first three steps taken to
  clean the data. The last element will report the data containing repeated
  measurements.
}
\description{
If the unit of BG measurement is mmol/L, this function, performs
  the following steps:
\itemize{
 \item{"<0.6" or ">33.3" reset to 0.6 and 33.3 respectively,}
 \item{"<" or ">" for middle values, decrement or increment by 0.1,}
 \item{"Low/lo" or "H/High" reset to limiting values 0.6 and 33.3,}
 \item{take the average BG at the same timing for particular patient, this
   might be system error,}
}
and finally reports affected proportions of readings.
}
\examples{
# Use a part of the example data for illustration:
data("gluDat")
gluDat <- gluDat[sort.list(gluDat$ADMISSION.ID), ]
dat <- gluDat[1:20, ]
# First format RESULT.DATE column in dat:
dat <- FormatDate(dat = dat, yy = 2016, mm = 7)
# Then distort some of the entries in dat for illustrative purpose:
dat$RESULT.DATE[2] <- dat$RESULT.DATE[1] # Make two timings identical for this stay
dat$RESULT[c(3, 8, 12, 15, 20)] <- c("<9.4", ">33.3", "c6.4", "lo", "<0.6")
# Make entry 12 unrecognisable
dat
# Now scrub dat:
DataScrubbing(dat = dat, unitVal = 1)
}
\author{
Ying Chen
}
