% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEpisode.R
\name{GenEpisode}
\alias{GenEpisode}
\title{Calculate pseudo admission episodes}
\usage{
GenEpisode(dat, epiMethod = "Pseudo")
}
\arguments{
\item{dat}{A \emph{data.frame}, normally is the outcome of function \emph{cleandata}, at least have three key variables:
\itemize{
 \item{RESULT.DATE} an object of classes \dQuote{POSIXlt} and \dQuote{POSIXct} representing calender date and times;
 \item{ADMISSION.ID} an object of classes "charater" representing the identification number for each hospital stay;
\item{RESULT} an object of classes "numeric" representing the Blood Glucose readings.
}}

\item{epiMethod}{This indicates the method takes to compute episodes. 
If None, the admission id will be assumed to be case id; 
if Pseudo, then episode number will be generated using 48 hours as a cutoff.}
}
\value{
A \emph{data.table} object, below are the variables that have been created:
\item{LOS.EACH}{time difference between readings}
\item{LOS.PSUM}{cumulative time differences between readings}
\item{EPISODE.ID}{In the case of pseudo, admission times will be increase by 1 if LOS.EACH greater than 48 hours. Otherwise, it will always be 1, corresponding to that each admission id identifies one hospital stay.}
\item{mond}{day of month}
\item{weekd}{day of week}
\item{yday}{day of year}
\item{hour}{numerical hours}
\item{weekn}{number of weeks in a year}
These variables are created using original key variable \code{RESULT.DATE}.
}
\description{
Read in glucose data and calculate pseudo admission episodes, calculate the time interval between consecutive readings
    and accumulated time intervals within each episode for each individual, identify the day in a week,
    month, year, week number in a year, convert timings into hours using the standard 24-hour clock.
}
\details{
none
}
\author{
Chuen Seng Tan, Ying Chen
}
