% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGluM.R
\name{GenGluM}
\alias{GenGluM}
\title{Generate glucometrics}
\usage{
GenGluM(dat, hypocutoffs, hypercutoffs, normalrange, hgicutoff, unitVal)
}
\arguments{
\item{dat}{A \emph{data.table} prepared by \code{\link{GenEpisode}}.}

\item{hypocutoffs}{A vector of numeric values indicating the recommended
cutoffs for hypoglycemia, from mild, moderate to severe hypoglycemia.
Hypoglycemia is defined as less than the specified cutoff value.}

\item{hypercutoffs}{A vector of numeric values indicating the recommended
cutoffs for hyperglycemia, from mild, moderate to severe hyperglycemia.
Hyperglycemia is defined as no less than the specified cutoff value.}

\item{normalrange}{A vector of numeric values indicating the recommended range
for normal glycemia, where the first value is the lower range, and the
second value is the upper range. A BG is considered in recommended range if
it is no less than a lower bound and less than a upper bound at the same
time.}

\item{hgicutoff}{A numeric value indicating the cutoff used for calculating
hyperglycemic index (HGI).}

\item{unitVal}{A unit indicator. 1 stands for mmol/L, 2 stands for md/dL}
}
\value{
Returns a list of three data.frames:
\itemize{
  \item{\code{popstat}:} A vector of all the indices at population level
  \item{\code{patientdaystat}:} A vector of all the indices at patient day level
  \item{\code{peradmissionstat}:} A vector of all the indices at admission level
}
which will be used as input to generate glucometrics tables.
}
\description{
Generate glucometrics for specific ward/wards during specified
  time period
}
\examples{
# Focus on data from Ward A. First prepare example data using GenEpisode:
data("gluDat")
gluDat2 <- FormatDate(dat = gluDat[gluDat$LOCATION == "A", ], yy = 2020, mm = 7)
gluDat3 <- GenEpisode(dat = gluDat2, epiMethod = "Admininfo")
# Then generate glucometrics:
metricList <- GenGluM(dat = gluDat3, hypocutoffs = c(4, 3, 2.5),
                      hypercutoffs = c(14, 20, 24), normalrange = c(4, 10),
                      hgicutoff = 10, unitVal = 1)
# View glucometrics (round to 1 decimal place):
lapply(metricList, function(m) round(m, 1))
}
\seealso{
\code{\link{ProGluTable}}
}
\author{
Ying Chen, Chuen Seng Tan
}
