% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenHypoIntervals.R
\name{GenHypoIntervals}
\alias{GenHypoIntervals}
\title{Calculate hypoglycemia management time intervals}
\usage{
GenHypoIntervals(dat)
}
\arguments{
\item{dat}{A \emph{data.table} prepared by \code{\link{GenEpisode}}.}
}
\value{
A \emph{list} containing the time intervals.
}
\description{
There are three types of intervals: hypo to next documented BG,
  hypo to resolution of hypo, and resolution of hypo to next documented hypo.
  This function will calculate these intervals.
}
\examples{
# Focus on data from Ward A. First prepare example data using GenEpisode:
data("gluDat")
gluDat2 <- FormatDate(dat = gluDat[gluDat$LOCATION == "A", ], yy = 2020, mm = 7)
gluDat3 <- GenEpisode(dat = gluDat2, epiMethod = "Admininfo")
# Then generate hypo intervals:
GenHypoIntervals(dat = gluDat3)
}
\author{
Ying Chen
}
