% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenGluTable.R
\name{ProGluTable}
\alias{ProGluTable}
\title{Produce Glucometrics table}
\usage{
ProGluTable(metricList, unitVal)
}
\arguments{
\item{metricList}{A list of glucometrics, from \code{\link{GenGluM}}}

\item{unitVal}{A unit indicator. 1 stands for mmol/L, 2 stands for md/dL}
}
\description{
Mainly used to produce glucometrics table
}
\author{
Ying Chen, Mark Salloway
}
