% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recogniseDateTime.R
\name{recog_one_time}
\alias{recog_one_time}
\title{Recognise one time (input as a vector if AM/PM is specified)}
\usage{
recog_one_time(time)
}
\arguments{
\item{time}{a string vector, where the second component (if any) indicates AM
or PM.}
}
\value{
Returns the time as a character variable of the format %H:%M:%S.
}
\description{
Recognise one time (input as a vector if AM/PM is specified)
}
