% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatDate.R
\name{recogniseDate}
\alias{recogniseDate}
\title{Recognise dates}
\usage{
recogniseDate(date, month, year)
}
\arguments{
\item{date}{a string vector of dates.}

\item{month}{A number (1 to 12) indicating the month this time point is in.}

\item{year}{A number (with 4 digits) indicating the year this time point is in.}
}
\description{
An alias to \code{recogniseDateTime} when time is not available.
  Uses the default \code{time_dummy} parameter in \code{recogniseDateTime}.
}
\examples{
recogniseDate(c("02/04/2020", "04/18/20"), month = 4, year = 2020)
}
