% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recogniseDateTime.R
\name{recogniseDateTime}
\alias{recogniseDateTime}
\title{Recognise date and time}
\usage{
recogniseDateTime(dateTime, month, year)
}
\arguments{
\item{dateTime}{A single time point satisfying the assumptions, see details.}

\item{month}{A number (1 to 12) indicating the month this time point is in}

\item{year}{A number (00 to 99) indicating the year this time point is in}
}
\value{
Returns a character variable of the format %d/%m/%y %H:%M:%S
}
\description{
Recognise date and time when month and year are known (otherwise
  there is no way to differentiate date and month)
}
\details{
Assumptions made on the format of input time are:
- only "/" or "-" is used to separate date, month and year, and no mixture
 of the two symbols in a single time variable
- only ":" is used to separate hour, minute and second, and the order is
  always in "h:m:s" (second can be missing, but not minute or hour)
- a dateTime variable should contain both date and time
}
\author{
Yilin Ning, Ying Chen
}
